polyjit: migrate run_with_likwid out of the experiment class
Use the static_var decorator to provide the correct configuration to the method
before using the cloudpickle.

This makes the code a little bit cleaner and explicitly states what we need from
our environment, without resorting to extra function parameters. In this case it
is actually cleaner, because we don't have to push them down the pipeline until
the project binary gets wrapped.


Former-commit-id: da43a6e81a70d02a38597fe307b1c07c8c33e5aa
Former-commit-id: 0e42d9ba05c149cd2262bec8007199481bd95158 [formerly 4e127f8034e6a0acd0658c0326ff4525ba0260dd] [formerly 57e0889d008bd06faa45b916000a3554d1a10335 [formerly 03d165606b4c9007def98f7e037525a56b74d4f0]]
Former-commit-id: 2d8d773ead9dd35588ed7d49098c99f9c093dc06 [formerly 8db8910d1128410cd682d6050c8cde4f4ea18516]
Former-commit-id: deaece1f0b3346c9e493b48b6faed24f30b49c2d