polyjit: migrate run_with_likwid out of the experiment class
Use the static_var decorator to provide the correct configuration to the method
before using the cloudpickle.

This makes the code a little bit cleaner and explicitly states what we need from
our environment, without resorting to extra function parameters. In this case it
is actually cleaner, because we don't have to push them down the pipeline until
the project binary gets wrapped.