add experiment_description to pprof runs
This adds an experiment_description to each pprof experiment run. By default
we just set the time of start for this experiment run.

In the case where this experiment gets executed on a distributed machine, the
last one to update the database config wins. Though, in general you can set any
description you like.

The buildbot will set the buildername-buildernumber as description for further
reference.


Former-commit-id: b90ff1019ef99c58c39822df8d1cc8aebea70562
Former-commit-id: d91364edfd7b9fbcd6072f5ea359e5a0589b98b6 [formerly f6a7d59dca0b37b3b94b830fbe998089276cefd9] [formerly f2d3a86ed198535bec2d76cc9f4b8d43f3763ab5 [formerly baf4e15edea59309f7ae0236614d0ca3a7f82279]]
Former-commit-id: 4028295ae4409dbbddb0af53b982f80905e93dec [formerly d48e8f38c7db9810660b5203ffb6f9c3ea6a8e2d]
Former-commit-id: 1e02817cb5c698612445e00283cc85f4e74b9c69