add experiment_description to pprof runs
This adds an experiment_description to each pprof experiment run. By default
we just set the time of start for this experiment run.

In the case where this experiment gets executed on a distributed machine, the
last one to update the database config wins. Though, in general you can set any
description you like.

The buildbot will set the buildername-buildernumber as description for further
reference.