remove persist_globalconfig
Temporarily remove persist_globalconfig from use, but keep it in the schema.

We need to fix the duplicate key error first, when executing on the cluster.
This is the result of passing the experiment ID via environment variables.

After we reintroduce this change, make sure we check for the existence of the
global config first.