remove persist_globalconfig
Temporarily remove persist_globalconfig from use, but keep it in the schema.

We need to fix the duplicate key error first, when executing on the cluster.
This is the result of passing the experiment ID via environment variables.

After we reintroduce this change, make sure we check for the existence of the
global config first.


Former-commit-id: 0b5703c3e3e3c044191be0b78de493673daabbb6
Former-commit-id: 88ec78ab9580229642b44ade897717ac2ad68a37 [formerly 72c73d3835590e44e6b356c2bbd79775188dc326] [formerly 4a4af9a26b1284706138b5b8aef5f0df7ed8fa71 [formerly 031766077f88bc4de43403212fd3c61fbe430ed5]]
Former-commit-id: 0226094d1585a28eb8f8497763e4bc68c66ada7f [formerly 943dba7e49b92062924eb8aaa066a23b0c23f2fa]
Former-commit-id: 20413be9d555ffdf6bbbd7ffa465e52e0f0146b5