gentoo: Add PrepareStage3 project
This project uses the existing stage3 archive for all gentoo projects
and spawns a bash inside. You can install all necessary prerequisites for
your measurement environment inside there. As soon as you exit the environment,
a new stage3 tarball for gentoo will be built and placed in the srcdir.

The next run with the gentoo project will automatically use the new stage3
tarball. You can repeat this process as often as you like.


Former-commit-id: da5bf50bd0f79d7206e1b82d59bfc5e9b667fd9f
Former-commit-id: 5ab1bd483573d25385e861056ee7b2d16d39f5f8 [formerly e5ef413549634147be4fc617c315dcb38c0d1fa4] [formerly 8c70ec6eb6dd8f2173f4442d2b83f3806612ab2e [formerly 1b65af27a1e5eaa91e40ab4a79dbcd44d302e580]]
Former-commit-id: fe675e29c41adb5fe1ac829650e4daff68b393df [formerly c5dd582d3dd3e92881e159821c7f22e06eb86daa]
Former-commit-id: 1c91fa71c1408ded9620f6ede66e484c2f9dc4bf