gentoo: Add PrepareStage3 project
This project uses the existing stage3 archive for all gentoo projects
and spawns a bash inside. You can install all necessary prerequisites for
your measurement environment inside there. As soon as you exit the environment,
a new stage3 tarball for gentoo will be built and placed in the srcdir.

The next run with the gentoo project will automatically use the new stage3
tarball. You can repeat this process as often as you like.