settings: prepare a new settings class
This will replace the current settings implementation.
We will go towards a decentralized declaration of configuration options.
The idea will be that a tool/experiment that is implemented inside the
pprof study can define a set of configuration options close to the point
where they make the most sense.

The new configuration will be stored in the JSON format outside of pprof-study.


Former-commit-id: 3e6e5f106888d61a2cad5ebcb9aa8ebeecf76c83
Former-commit-id: d228aa56ee11f19c550a5ac07189df8cd852bcea [formerly f690415c3c70756b7479b2a7bb9e5c60e325c392] [formerly 5e455ce2bbe00a96063ade56f827898e11005a97 [formerly a375119b90e71f1f2bd209d8cf24d199be337bce]]
Former-commit-id: c2c51317e57da79b40d2779f3d4f8ab849622b6c [formerly 24008f241889771da1aa574c652577680bbebfb9]
Former-commit-id: ddd406dd7aa3c886c8c73ad93c04d003c210c4d4