settings: prepare a new settings class
This will replace the current settings implementation.
We will go towards a decentralized declaration of configuration options.
The idea will be that a tool/experiment that is implemented inside the
pprof study can define a set of configuration options close to the point
where they make the most sense.

The new configuration will be stored in the JSON format outside of pprof-study.