gentoo: use memento function to cache result of curl call
The new @cached decorator can be used to make a function
remember its return value. Repeated calls to a cached function
will not execute its side-effects, but only return the cached
result.

In this patch we use it to avoid calling curl on gentoo's servers
multiple times per project run.