gentoo: use memento function to cache result of curl call
The new @cached decorator can be used to make a function
remember its return value. Repeated calls to a cached function
will not execute its side-effects, but only return the cached
result.

In this patch we use it to avoid calling curl on gentoo's servers
multiple times per project run.


Former-commit-id: 00e0b844e69beffdea61da7a18a39b7205b92a20
Former-commit-id: f26d6a56d2f4d1da5ced033c748266c1e7511674 [formerly ad3a6b66971488c1fb3dcb0f4922e7a7dc4d3a33] [formerly 72b93c0ed6368e6a32bc238868946a7e6e41ae93 [formerly 39e8b83226af7497788eb596b28e22e51645c3d9]]
Former-commit-id: bdeaaa38a17684b77612f2fb740766f51939f03e [formerly 14a0c4e237112f78e7c10f391d1a136cdcc06ba6]
Former-commit-id: 5e1192ce9c2e379476f437963732c5e1e215e0c2