portage_gen: add reducer, disable run in dynamic Portage Projects
This disables the run() method in portage projects. There are no unit tests
anyway, so we make sure that we never trigger the default implementation from
<pprof.project.Project>.

Furthermore, in order to support pickling in compilestats experiments, we add a
custom reducer that advises the pickle(dill) module in pickling/unpickling a
dynamic portage project. Basically, we just state how to instantiate the object
with the same attributes as before.

This is in no way 'safe' in terms of preserving the internal state of the object
yet.