portage_gen: add reducer, disable run in dynamic Portage Projects
This disables the run() method in portage projects. There are no unit tests
anyway, so we make sure that we never trigger the default implementation from
<pprof.project.Project>.

Furthermore, in order to support pickling in compilestats experiments, we add a
custom reducer that advises the pickle(dill) module in pickling/unpickling a
dynamic portage project. Basically, we just state how to instantiate the object
with the same attributes as before.

This is in no way 'safe' in terms of preserving the internal state of the object
yet.


Former-commit-id: fec1c0f0dc4bf915e937164edb3abe1f6678fdee
Former-commit-id: fd95964a6e8c132b8b4719c8139d756bc09399e8 [formerly e93800eef453ba817df4fce212e1d04bfe6fa5f0] [formerly 526a7921c6944a62289a726ae59ae408e987ca77 [formerly 04f798aa91adf3249462c49e9af95d5807064bc4]]
Former-commit-id: 0cc972f74df185b8436748476147e17c01b93ee7 [formerly c6acb82f51f48f30e2add7775b97dd5d04f95586]
Former-commit-id: b61ccf2206983596150fbfea27cbabb24755f357