container: add a container management driver
This commit adds a separate 'binary' that takes over the role of
the old gentoo stage3 project.

With calls to container you can setup a fresh container from any
linux base image and pack it into one that can be used with
benchbuild.

The output of container help gives more information about the
available flags.