container: add a container management driver
This commit adds a separate 'binary' that takes over the role of
the old gentoo stage3 project.

With calls to container you can setup a fresh container from any
linux base image and pack it into one that can be used with
benchbuild.

The output of container help gives more information about the
available flags.


Former-commit-id: 55941daac97220711f5578349da79694c480badb
Former-commit-id: 50fd33da6db321561fd27103d512a2a19db405bb [formerly d009e1212ca23d764a95f10d62d79fa7e2def419] [formerly b638a9675be7724244a413ce631e1eb837433481 [formerly e13352584280d4645a17ef3091a1b2a634019e18]]
Former-commit-id: 10cca9066dfa329258ea87f005374328af38670b [formerly 59903c5f2b29b59c1d9cf69e3f283ba59f5429c6]
Former-commit-id: 6620979b7f2df49c804012a68fe1602926791068