project: add automatic decoration of the project interface
This adds a new metaclass to projects that decorates all interface methods.

At the moment we just decorate the methods with an
in_builddir decorator.
This decorator ensures that all plumbum commands are executed
within the project's build directory. This provides a small safety net
for users, because benchbuild guarantees that the current working directory
inside a project is always it's build directory.

The rest of the commit just refactors all of benchbuild's projects to make
use of that feature.