project: add automatic decoration of the project interface
This adds a new metaclass to projects that decorates all interface methods.

At the moment we just decorate the methods with an
in_builddir decorator.
This decorator ensures that all plumbum commands are executed
within the project's build directory. This provides a small safety net
for users, because benchbuild guarantees that the current working directory
inside a project is always it's build directory.

The rest of the commit just refactors all of benchbuild's projects to make
use of that feature.


Former-commit-id: b2449f789c207676fd439db1f848836ee48f288f
Former-commit-id: e8c6fe052671bc31cd279503be94a319a4d74f1b [formerly f3eef79966ee242f0c3dc57776700e02bb367873] [formerly 372e231acc30423b26197ef97bc63a55045303b5 [formerly 504ba121bc59a7079d170b7c0728257e513309a1]]
Former-commit-id: c4cbfd842d70debdec2cbf1f50390282023c8b8e [formerly 4f06c06886484714a8d82a92d20a32794dc851a6]
Former-commit-id: e1253a98e69d9f3b816643a081d4c0921a436a7f