unionfs: add a decorator that redirects project execution into a unionfs
This adds unionfs support to benchbuild. 2 directories are stacked into a single
filesystem. The base_directory stays read-only and is 'below' the image_dir,
which catches all write-accesses.

The image_dir does not need to be inside the project's build directory (by
default it is).
The configuration options control unionfs behavior:
* CFG['unionfs']['image_prefix'] -> A prefix for the image dir, e.g.
    '/dev/shm'. Benchbuild will take care of cleaning the external directories.
* CFG['unionfs']['enable'] -> Enable/Disable unionfs. (Default: Enabled)