redirect plumbum.cmd to benchbuild.utils.cmd
This commit redirects all calls to our own utils.cmd module.
In there we check all requested binaries for available aliases.

In a first effort we use this to enable arbitrary aliases for the same command.
As an example:
unionfs-fuse defines a binary called 'unionfs-fuse' in version 0.24
   (Ubuntu Trusty),
but 'unionfs' in version 1.0 (Gentoo ~amd64).

This can be used later for dynamic wrapping in uchroot.