redirect plumbum.cmd to benchbuild.utils.cmd
This commit redirects all calls to our own utils.cmd module.
In there we check all requested binaries for available aliases.

In a first effort we use this to enable arbitrary aliases for the same command.
As an example:
unionfs-fuse defines a binary called 'unionfs-fuse' in version 0.24
   (Ubuntu Trusty),
but 'unionfs' in version 1.0 (Gentoo ~amd64).

This can be used later for dynamic wrapping in uchroot.


Former-commit-id: 86ca08d2fde9da3f922ba9dbd3ee6fb57e733e87
Former-commit-id: 706e94aa5188216b4b0427b4806c04e2fae1aa4d [formerly 6382ffbe086783831e13ee02d5113c9f3261a155] [formerly ca802f8394b4aa564717f38fb24abfe05b34b232 [formerly ae8a1c0190aed965059820ba867fd0df27a5be8c]]
Former-commit-id: 1a18ea27f0b5091becd3ef1e39a47f4001a0a1de [formerly 9c46fcaf667f708dea9f8b2160db19e8ca147776]
Former-commit-id: 989290bfe00b37900f511e71ae3feaebd006c245