fix: create files inside the uchroot container
Without this, we would write to crucial configuration files from outside of the container. When subuid/subgid emulation is enabled in uchroot, this leads to wrong permissions inside the container.

With this commit we first create them via 'touch' or 'mkdir' from the inside and write to them as usual form the outside. This way, our rights should be preserved.

Former-commit-id: 59f64d25578fce2004008e4aff06958b1b5c9cea
Former-commit-id: 8398397330568a175bc4252a101b7c13759b7608 [formerly fecbd07b36a2856ef8d09d1f6f262791e2d988de] [formerly 78a98c2151bef922651356092e6179532e2d8de7 [formerly 5b60e3e1f28b8e423053a54a35a5c74a5e30d1f3]]
Former-commit-id: 480f13188c25992b1a64df6e36ae7d843d19f379 [formerly c0b5697bf167ae8b2047971d0bb3df199cbdb853]
Former-commit-id: 4f2b544dbad491bc5795dd1d1fa0d115fdaa7bac