fix: create files inside the uchroot container
Without this, we would write to crucial configuration files from outside of the container. When subuid/subgid emulation is enabled in uchroot, this leads to wrong permissions inside the container.

With this commit we first create them via 'touch' or 'mkdir' from the inside and write to them as usual form the outside. This way, our rights should be preserved.