started working on the versionizing of projects
Added variable VERSION and created a function that finds out the hash
of a project before initializing it by working with the source file.
Therefor every project must offer a source file and a SRC_FILE variable
was created in the project class as well.
Also warnings were added if a project is initialized without a version
or a source file.