started working on the versionizing of projects
Added variable VERSION and created a function that finds out the hash
of a project before initializing it by working with the source file.
Therefor every project must offer a source file and a SRC_FILE variable
was created in the project class as well.
Also warnings were added if a project is initialized without a version
or a source file.


Former-commit-id: 3c8808678b92d9bdb62d7548650acd9e45779639
Former-commit-id: b1f8a0a32016bed2fba17ab8c84fb43803766648 [formerly 81777d9177034e50ab126a64bea50bb539e3425a] [formerly d93eba9dfab3cf7962c993c49c1164e5d7bf0194 [formerly ed711d97990415cc03f8e94df70149b3251d89cf]]
Former-commit-id: 71dad5456581419aa8096f6c6a90f0227b35fe09 [formerly 10b2183e6f757336348a4aa86ed5b64668861fae]
Former-commit-id: 98645fdbb605a13d713556178c74bb320af1d426