version: refine version calculation
This commit refines the existing version calculation. It tries to minimize the number of calls to get_hash_of_directories because this call is very expensive on slow disks (NFS mounts).
There are many more situations where we do not need to rely on the has calculation because it might have already been done once (.hash files).