container: introduce new strategy based container creation
This adds the possibility to execute an arbitrary strategy for
container creation. We implement 2 simple strategies:
1. the 'bash' strategy, which spawns a bash inside the container.
2. the 'polyjit' strategy, which requires a gentoo container and
     installs all dependencies necessary for running polyjit
     inside a gentoo container.

This implementation is very 'raw'. The 'polyjit' strategy does not
work yet. We still need to port image generation from the gentoo projects
completely.
When this is finished, we can remove the gentoo 'meta' projects.