container: let gentoo merge an arbitrary number of packages
This makes the gentoo strategy more flexible from a user perspective.
New package dependencies do not require code changes at all.
You can control the packages that are merged inside the container with
a dictionary in the benchbuild configuration

The node CFG["container"]["strategy"]["polyjit"]["packages"]
gives you a list of dictionaries of the format
{"name": <package-atom>, "use": [<list-of-useflags>]}
which allows you to customize the configuration and the version of each
package.