container: let gentoo merge an arbitrary number of packages
This makes the gentoo strategy more flexible from a user perspective.
New package dependencies do not require code changes at all.
You can control the packages that are merged inside the container with
a dictionary in the benchbuild configuration

The node CFG["container"]["strategy"]["polyjit"]["packages"]
gives you a list of dictionaries of the format
{"name": <package-atom>, "use": [<list-of-useflags>]}
which allows you to customize the configuration and the version of each
package.


Former-commit-id: 1c22a82677cce423bebc7e5b07df0a5dac797ed7
Former-commit-id: f581cb9ffa0fbe5d4aa4235676996d71d8987f20 [formerly c2cc6ef5662a904c61413ce57618539569f61e82] [formerly 9ced7a1b44748a7440815b8464b480bd5559cce9 [formerly 8b04212ad49bce36313dd99cd3bf643c232d4e0c]]
Former-commit-id: 5883de9ebc47ca7f0559e45b60679848c69941d2 [formerly ec108d10da4645bc5a12e7fb8dc0023d192dce44]
Former-commit-id: 54e3a057152df699af817d7b977b1f070541012c