project: store_config _after_ unionfs is set up.
We need to change the order in which the project API decorators are
applied.

On the outermost level we need to set up the unionfs filesystem,
if requested by the user. The unionfs reconfigures parts of the
benchbuild configuration and, therefore, we need to make sure that
all subsequent decorators operate on the latest configuration data
structure.

For now we have the order:
Unionfs -> Store_Config -> In_Builddir


Former-commit-id: 77c0126df880c3f695284812fcd772abff5396e5
Former-commit-id: 6de03a930d8e33411ed0d14fc71999bb9f4056ed [formerly f5f9c1a74c812c1f38e125d1f5312c20a3ac8b04] [formerly 5abc52c49df7ba90527222fbc9c796a53745d014 [formerly 466e112518a5e91cd371f6981c5cdf5c6ee939c6]]
Former-commit-id: c975ec54a09560a4b2535dbb83f364ad61bdd53f [formerly 79f1ea3432644832697ff2712c83caf7582c929b]
Former-commit-id: 9e23945025a2345f2c731ea889dfa8438e0d22eb