project: store_config _after_ unionfs is set up.
We need to change the order in which the project API decorators are
applied.

On the outermost level we need to set up the unionfs filesystem,
if requested by the user. The unionfs reconfigures parts of the
benchbuild configuration and, therefore, we need to make sure that
all subsequent decorators operate on the latest configuration data
structure.

For now we have the order:
Unionfs -> Store_Config -> In_Builddir