actions: make Clean action aware of UnionFS
This is required in the case something unforseen happens
to our UnionFS mounts (Machine-Crash/ Job Abort, without time to
cleanup).

Before we clean the desired directory, we check for a mount
with the type 'fuse.unionfs' below or at the desired directory.
All mounts we could find are added to a worklist that will
be fed to unionfs_tear_down() one by one afterwards.