actions: make Clean action aware of UnionFS
This is required in the case something unforseen happens
to our UnionFS mounts (Machine-Crash/ Job Abort, without time to
cleanup).

Before we clean the desired directory, we check for a mount
with the type 'fuse.unionfs' below or at the desired directory.
All mounts we could find are added to a worklist that will
be fed to unionfs_tear_down() one by one afterwards.


Former-commit-id: 1d30b644fd42afc93296fde3fd95898da47e5f15
Former-commit-id: 37796c999e8eb694fd7fb2fbab4a2a4b0dbbeaf5 [formerly 18b8d260a54fb870eeefd9e6bfc3e30e9e0de221] [formerly 4c134c4ad851168b5fc5e2356decaa8f5055417e [formerly c1c1a720eaddb1b6adc78e80c966a0dfc063c0b7]]
Former-commit-id: 12623095d1f3e00219c03ab0fe349192dcf789ca [formerly 6038af7e60136b611f1445224cc719ebfb87ad88]
Former-commit-id: 9b230f89015e556218bbbac66b9057299d98422f