pj_sequence: get rid of multiprocessing module
We use concurrent.futures instead and work with a standard thread pool
instead of a process pool.

In the current implementation we use the number of jobs available to
us as configured by the user. All jobs are spawned concurrently for
each element of the base_sequence (prepend/append one) and
the results are gathered aterwards as they complete.


Former-commit-id: 909e6d8bd16359db39487029dcb8f934ea406611
Former-commit-id: 847330ca50dc285f92ffbd06a5e03845e96c9afe [formerly 7b926dbedfab43312e8554184238391679de2fb5] [formerly 47059374af02e087e87d43f2474ff9afd280c48b [formerly b588add82320a9e4d03816eb0e2c5e9eddcf74c9]]
Former-commit-id: 92e352be435aa3fdfa8d558c8600269818952f38 [formerly c4d12ddf4cac6ab42d7ba1dd8b5acbc2e32f663c]
Former-commit-id: f93ff1a23902913f6c58d62328cf566a2737bac0