pj_sequence: get rid of multiprocessing module
We use concurrent.futures instead and work with a standard thread pool
instead of a process pool.

In the current implementation we use the number of jobs available to
us as configured by the user. All jobs are spawned concurrently for
each element of the base_sequence (prepend/append one) and
the results are gathered aterwards as they complete.