pj_seq: Refactor the genetic algorithms
The simulation of multiple generations was a bit of a monster-function
and i transfered the crossover of two genes into an extra function.
This way it is also more easy to see how the two genetic algorithms
differ from each other, since the difference is within the new crossover
function.


Former-commit-id: bce26f49326dd4512d468c083818b77a03282dad
Former-commit-id: d8cae172429912f64eb62b6a7e5a560e371214aa [formerly c48cc4ada764cd88a3a2e4a99207f5d367bdb740] [formerly 4540362cc855d368e23da4e39b8af32a7c317748 [formerly c9788e0536d53b601df6b28bbba0c51d26508667]]
Former-commit-id: 239dd0d5e25cd991ddc301414c975afc357709a1 [formerly 6cd82a26745fe4697d43b0e7e4f251a651150a32]
Former-commit-id: 4b4cde310ac7da64fa065b03c5e8734382734ef3