pj_seq: Refactor the genetic algorithms
The simulation of multiple generations was a bit of a monster-function
and i transfered the crossover of two genes into an extra function.
This way it is also more easy to see how the two genetic algorithms
differ from each other, since the difference is within the new crossover
function.