container: add prefix support to containers
This commit adds support for internal prefix directories inside the
container. These are treated the same as external mountpoints for
containers, i.e., ./bin and ./lib subdirs are added to PATH and
LD_LIBRARY_PATH.

How this is done is depending on the project that uses the container.
Gentoo projects modify the bash environment accordingly.