container: add prefix support to containers
This commit adds support for internal prefix directories inside the
container. These are treated the same as external mountpoints for
containers, i.e., ./bin and ./lib subdirs are added to PATH and
LD_LIBRARY_PATH.

How this is done is depending on the project that uses the container.
Gentoo projects modify the bash environment accordingly.


Former-commit-id: 0d9ddc6d42a158c0a67b543e7f0acb00adb23b68
Former-commit-id: 2bdbdf323c455942dc4e15ff83d060503a50814e [formerly 7de363ecf967da61a412cd74db051cb220cb39fb] [formerly 5cc987dc52f013f4686e41e43a4e511d4ccce217 [formerly 9bcdca9bd00c4822068567ea0f5339d2f2565c71]]
Former-commit-id: 2837f442993aa9517f995808ffb59e155b9ddb7e [formerly 6aa146d36a9ebeb83efc0f6a5e898b3bd5896a88]
Former-commit-id: e45bfb3a75539e12fd8a96875640d78bf222afc0