dict: add a new (extensible) dictionary
Our ExtensibleDict can be used similar to plumbum's context managers
that modify the environment of a program (stack-like).
However, we can append/extend the entries of the environment in contrast
to plumbum's 'overwriting' functionality. This enables us to expand the configuration of a program step-by-step and construct the final configuration
as late as possible.


Former-commit-id: 606ef879a331adbdd5c213c2d539caf12f4ddc31
Former-commit-id: cd41214e388da9779953ce502b018d70d5298023 [formerly 399bbfc1ec7e1243dc28669a79af9b2da917479d] [formerly 2ec56cc5d720f1443127a3de70f3f280aba7b3a8 [formerly 9f8736e6e3d1d7431a096b52fd70f81f2b82c104]]
Former-commit-id: 4dede60d4593f0a7be11476ab8c91d27468ab336 [formerly 05ba53357b690ce4d1737434304cc810f24c86dc]
Former-commit-id: b4542a3e318bcd197d6eb7c38d9edd360c31d21c