dict: add a new (extensible) dictionary
Our ExtensibleDict can be used similar to plumbum's context managers
that modify the environment of a program (stack-like).
However, we can append/extend the entries of the environment in contrast
to plumbum's 'overwriting' functionality. This enables us to expand the configuration of a program step-by-step and construct the final configuration
as late as possible.