schema: do not allow nested sessions by accident
This fixes a severe DB management error. Since introduction of the new
extensions API it was possible to create nested Session objects. These do not
actually get committed as the user expects.

We now do not allow more than one session to be created during one benchbuild
run. This avoids this problem automatically.


Former-commit-id: dfd63b5752c75021fb648deb816557a8fc8a3e4b
Former-commit-id: ae447f34f699f024ca470adc17f4b96b84668e09 [formerly 1d21a9d681e61adc36f5e035fb3de45e63c6fa58] [formerly a2a3f53acff53845251ff377ad0b502035bdf4e1 [formerly 80d1d9fef8055a2165381fe6960cbadcff3780d8]]
Former-commit-id: 12828c29bf0622857bd9da5bc8f83e8bb558ffba [formerly 99e427185626bebe54ca2dac9086e7e452fc3daf]
Former-commit-id: 5845367583f4d922606e1a2e31e8cefda1b0c87b