schema: do not allow nested sessions by accident
This fixes a severe DB management error. Since introduction of the new
extensions API it was possible to create nested Session objects. These do not
actually get committed as the user expects.

We now do not allow more than one session to be created during one benchbuild
run. This avoids this problem automatically.