replace experiment_id config options with dictionary
This replaces the old experiment_id uuid configuration options to a
complete dictionary. The goal of this is that we want to be able to
run multiple experiments under distinct experiment id's. For this we
actually cannot rely on a single uuid in the config for all our
experiments.

This is enabled for normal binary runs but not yet fully supported
for slurm runs (the slurm tool is not yet able to generate scripts
for multiple experiments).