replace experiment_id config options with dictionary
This replaces the old experiment_id uuid configuration options to a
complete dictionary. The goal of this is that we want to be able to
run multiple experiments under distinct experiment id's. For this we
actually cannot rely on a single uuid in the config for all our
experiments.

This is enabled for normal binary runs but not yet fully supported
for slurm runs (the slurm tool is not yet able to generate scripts
for multiple experiments).


Former-commit-id: bea82fcab20203254e8a3979ad3402208c1a583d
Former-commit-id: 123212e00650e4fc477118e6724947a7dfc6b271 [formerly d0471ae6463799b5bf1cef3c44ab012aae2fc3eb] [formerly 9e13076a7480fb8b792bef7ae765fd81ead3c9b0 [formerly 7df39b417d2a28ba0535957c7663ab1c57df143c]]
Former-commit-id: a9d81c2bcb4bbfdc5ed4148854872136014b147e [formerly 67a1f0606a4cd5a53c4e61776a5e845925b9821e]
Former-commit-id: c0ba70315a3368bf35c24eeab8dac67f3ef6abbc