New Project: Barcelona OpenMP Task Suite
The objective of the suite is to provide a collection of applications
that allow to test OpenMP tasking implementations. Most of the kernels
come from existing ones from other projects. Each of them comes with
different implementations that allow to test different possibilities of
the OpenMP task model (task tiedness, cut-offs, single/multiple
generators, etc). It currently comes with the following kernels:

Alignment: Aligns sequences of proteins.
FFT: Computes a Fast Fourier Transformation.
Floorplan: Computes the optimal placement of cells in a floorplan.
Health: Simulates a country health system.
NQueens: Finds solutions of the N Queens problem.
Sort: Uses a mixture of sorting algorithms to sort a vector.
SparseLU: Computes the LU factorization of a sparse matrix.
Strassen: Computes a matrix multiply with Strassen's method.

Former-commit-id: efd9b27ceb8cfef819a1ea1f32868addbfd5f3c3
Former-commit-id: 12774aef72ebb4e1ac5d33dafc18c0a66644d6d1 [formerly 29fed246dcfeeddf391f44e74b0461e73277269c] [formerly 23b43dceebc7338b22ad4449f34118dc55b8c4cd [formerly 4bb9b279c92841a7dabe611ca04970af723ba61f]]
Former-commit-id: 9711e954b6386657659033b0c350e8cb942fb769 [formerly cb30870837a84a8007921368ea1b1bfd6edbb339]
Former-commit-id: f8c478df51d20834c94083e4430a2ff2c5d9f31e