New Project: Barcelona OpenMP Task Suite
The objective of the suite is to provide a collection of applications
that allow to test OpenMP tasking implementations. Most of the kernels
come from existing ones from other projects. Each of them comes with
different implementations that allow to test different possibilities of
the OpenMP task model (task tiedness, cut-offs, single/multiple
generators, etc). It currently comes with the following kernels:

Alignment: Aligns sequences of proteins.
FFT: Computes a Fast Fourier Transformation.
Floorplan: Computes the optimal placement of cells in a floorplan.
Health: Simulates a country health system.
NQueens: Finds solutions of the N Queens problem.
Sort: Uses a mixture of sorting algorithms to sort a vector.
SparseLU: Computes the LU factorization of a sparse matrix.
Strassen: Computes a matrix multiply with Strassen's method.