wrap_dynamic: add regex filter support for dynamic wrapper.
This adds a new parameter to wrap_dynamic, used by lnt based projects:
  SingleSourceBenchmarks
  MultiSourceBenchmarks
  MultiSourceApplications
  SPEC2006
  Povray (LNT)

You can now provide a list of regex strings to the function using the
named parameter 'name_filters'. The wrapper script will try to match
them in list order and select the match group that is named 'name' out
of it.