wrap_dynamic: add regex filter support for dynamic wrapper.
This adds a new parameter to wrap_dynamic, used by lnt based projects:
  SingleSourceBenchmarks
  MultiSourceBenchmarks
  MultiSourceApplications
  SPEC2006
  Povray (LNT)

You can now provide a list of regex strings to the function using the
named parameter 'name_filters'. The wrapper script will try to match
them in list order and select the match group that is named 'name' out
of it.


Former-commit-id: 85a5efec9766b81a1a4708d55ed73b45a4db69da
Former-commit-id: 6b63726a1f5badb272a7e38dfe2be3e647268f55 [formerly 618589a95d7c2886b83e0390e8b9566a584dadeb] [formerly ab48ee984235988293a2f24fb5aeea54b6bac6db [formerly 486c3882a25b5ae48e9a76111231911d5c17a383]]
Former-commit-id: 5f13750b8f7fb9362cecaf7a36c42f4b71a2307d [formerly 16c5c7e53447ccd99668189fde037caa2610473d]
Former-commit-id: e27daafd0f54b23b3b9e4014b820a6e8dc409215