Change the statistics extension to simply repeat an experiment several times
This tests the usage of the extension with the pollytest experiment.
From this point on i can design the interface later experiments should
use to calculate results out of a run info object generated by the run
of an experiment.
The statistic extension shall analyze the significance of those results.


Former-commit-id: 3525fb8c26516fba98589195228d9b2570595aae
Former-commit-id: bfe6ab876d0f0fa54b5d59e16b20462de8c0d061 [formerly b91dfea1d0be52255493fb1b24c965d5c714186f] [formerly 206462eb1ff469c3f118201f099d6d8bfcb6e2bb [formerly c7190bcef37e6bc69858ff72858abaad45f42d20]]
Former-commit-id: 277fa257c9492850802e99389420472bd98a9b7f [formerly a3860dc1ca63692603cb222a14fb6edbebea405f]
Former-commit-id: a227b3f0f47d3cd426fc84f66bc5161d1ed39a68