Change the statistics extension to simply repeat an experiment several times
This tests the usage of the extension with the pollytest experiment.
From this point on i can design the interface later experiments should
use to calculate results out of a run info object generated by the run
of an experiment.
The statistic extension shall analyze the significance of those results.