compiler: add configurable compilers (experimental)
This remove the hard dependency on clang/llvm for benchbuild.
We add a new configuration section to manage the known compilers.

By default we already support the following entries:

compiler:
  cc: default: clang
  cxx: default: clang++

We look for the compilers in the currently active PATH, make sure you have
the configuration section set up correctly.
env:
  path: ...
  ld_library_path: ...

Former-commit-id: ee4150203b25ebd1dffc6f6734dda424e0c31947
Former-commit-id: e428f0890014bf847ab907be9304d0ec7eef55b6 [formerly 54abae90de4d892b50267077141e0a69348abfa2] [formerly a6e641b273fb719a21d86e33bf5cabf2ef6e6f3d [formerly 3a5f3cbd90f11ad45ebb193d716d66a9f77dfc7e]]
Former-commit-id: 0572c6f59d1cb1ac11c2fe86552960e01ec356f2 [formerly 825d0311a9730a3230630e30bc9ab1e22d983fd3]
Former-commit-id: c6d2d3e01e47b18e3a61ff1d5b1cf9ec3856215e