compiler: add configurable compilers (experimental)
This remove the hard dependency on clang/llvm for benchbuild.
We add a new configuration section to manage the known compilers.

By default we already support the following entries:

compiler:
  cc: default: clang
  cxx: default: clang++

We look for the compilers in the currently active PATH, make sure you have
the configuration section set up correctly.
env:
  path: ...
  ld_library_path: ...