sqlalchemy: improve support for different database drivers (sqlite3)
This masks a few configuration options from sqlite3.
In general benchbuild already works with sqlite3.
However, a few race-conditions are already known, so do not expect
sqlite3 to be reliable in any way.

In this patch we switch the database configuration over to a monolithic
setting. Instead of configuring hostname, port, user, pass with their
own setting, we now provide only one option for the connect string.

db:
  connect_string: ...

Please refer to sqlalchemy's documentation on connect strings.