sqlalchemy: improve support for different database drivers (sqlite3)
This masks a few configuration options from sqlite3.
In general benchbuild already works with sqlite3.
However, a few race-conditions are already known, so do not expect
sqlite3 to be reliable in any way.

In this patch we switch the database configuration over to a monolithic
setting. Instead of configuring hostname, port, user, pass with their
own setting, we now provide only one option for the connect string.

db:
  connect_string: ...

Please refer to sqlalchemy's documentation on connect strings.


Former-commit-id: 9446d48cd55ad70855f73aa5cf532cbfb3fb1af0
Former-commit-id: 18c544d5c5bd9935aa1c43cc9d7812d0df8bb9af [formerly 53dfaf86c36a733a3c95c861d7006afb826ae072] [formerly f79a0706685433cba6dcd560986dd3842d87e430 [formerly 22658db83e47c68781ecbe16c5c2f9055f2aef63]]
Former-commit-id: 5c30cf3016ff45e87a3b6587c032acdb13ddeb68 [formerly 3849b2105d44248633d01ca7d77fd3f1eb1b4419]
Former-commit-id: 11871b0eca0f591ad647c430aa6e72c65e3231ba