schema: support automatic database upgrade using sqlalchemy-migrate
This enabled auto-upgrade support for sqlalchemy-migrate scripts.
We use the following workflow to perform database migration:

If the databse is not under version control -> Ask the user to enforce it.
  This step assumes that you are operating with the latest benchbuild
  schema before this commit.
If the databse is under version control compare version in the database
to the version in the repository.
  If the versions do not match, upgrade to latest.

We do not consider downgrades, but try to implement all upgrades bidirectional.