schema: support automatic database upgrade using sqlalchemy-migrate
This enabled auto-upgrade support for sqlalchemy-migrate scripts.
We use the following workflow to perform database migration:

If the databse is not under version control -> Ask the user to enforce it.
  This step assumes that you are operating with the latest benchbuild
  schema before this commit.
If the databse is under version control compare version in the database
to the version in the repository.
  If the versions do not match, upgrade to latest.

We do not consider downgrades, but try to implement all upgrades bidirectional.


Former-commit-id: 58abc4f22b3af8071d43ffc1ea368c822be041cc
Former-commit-id: 6fbda75e1dde79c67382cc84895f3ec83c4ba94d [formerly 9179cad30c70b53a454ae98af9d63160104b471e] [formerly a69616707b36bfc89feb04eb83a316265edca747 [formerly f4d0edeee7fd13df21de437bc6151a875c4d43d0]]
Former-commit-id: e7ec816b9cf0d2a04ba7e04a977de73420cb25fe [formerly 33acde23573aac5b34e896bece0e696c86c19198]
Former-commit-id: e44c5b1c18a9ebb1f3601bcf54bf559e24d32ff6