experiment: add schema support
This just adds a (for now) unused attribute that contains the 'schema'
associated with this experiment. The idea is to be able to reason about the data
we can expect from an executed epxeriment.
Here we distinguish between the schema generated by benchbuild (Run, RunGroup,
Experiment, Project,...) and the schema added by a given experiment.

We still spawn the complete schema on first contact with the database, so
nothing will change for the user. However, we can implement a targeted data-dump
once we have access to the subset of tables relevant to a given experiment.


Former-commit-id: fef6c05a08321074871fe48e479804306a236e68
Former-commit-id: 21e680b3d74cd8ba8d4ee01ac2dd6779c68a917c [formerly 0105b7ceaeb6a1b374558bad77999c2f82584015] [formerly e94e28175970da12f1b9b4c6d529aa14075d4b01 [formerly 8f762a8884c738e44605c30b85220d3879f32d1c]]
Former-commit-id: e9aa36712622e014e9dbc6e24566b69a2e4f1784 [formerly 11a64903c21e883403f4194236dcc19672f2b3c6]
Former-commit-id: 207de1c10774025ce72599f3f0ac258ca4b22056