experiment: add schema support
This just adds a (for now) unused attribute that contains the 'schema'
associated with this experiment. The idea is to be able to reason about the data
we can expect from an executed epxeriment.
Here we distinguish between the schema generated by benchbuild (Run, RunGroup,
Experiment, Project,...) and the schema added by a given experiment.

We still spawn the complete schema on first contact with the database, so
nothing will change for the user. However, we can implement a targeted data-dump
once we have access to the subset of tables relevant to a given experiment.