utils/settings: improve YAML support
This removes JSON handling routines from the configuration. They are unused and
unmaintained.
Furthermore, we introduce better support for UUID's and Path-Strings in our
configuration.

If you want a configuration path to be checked by benchbuild at the start, you
can use the YAML tag '!create-if-needed' as a prefix to your path.

Example:
	build_dir: !create-if-needed '/important-path'

If it does not exist and we are on a tty, benchbuild will ask if it should
create it for you.
If we are not on a tty, benchbuild will silently create it.

The other changes accross benchbuild are descendents from the configuration
changes, as we now require/can clean up usages of path strings.