utils/settings: improve YAML support
This removes JSON handling routines from the configuration. They are unused and
unmaintained.
Furthermore, we introduce better support for UUID's and Path-Strings in our
configuration.

If you want a configuration path to be checked by benchbuild at the start, you
can use the YAML tag '!create-if-needed' as a prefix to your path.

Example:
	build_dir: !create-if-needed '/important-path'

If it does not exist and we are on a tty, benchbuild will ask if it should
create it for you.
If we are not on a tty, benchbuild will silently create it.

The other changes accross benchbuild are descendents from the configuration
changes, as we now require/can clean up usages of path strings.


Former-commit-id: e1af236085d9534adc6834c2779204e4b800b07d
Former-commit-id: ef3767995c3f95ef7e88f1e0d5fb06407cbc016d [formerly 962376971d9025f33c7a3e8fefffdc5aa8b650b6] [formerly be27f3154ee3fc9acc996b0b7502ab29c6c0c20e [formerly d3cf928e6102ebe38ab9b49428b029db57843951]]
Former-commit-id: 36eed1a06a95fe25e75d727d1aed29d587fb3751 [formerly 8254295a55437b4a2f54938bc907711f9145988f]
Former-commit-id: c4cdbc107470399a93aae1894fed20c39a83b278