extensions: add payload support
This adds support for custom payload to communicate run results between
extensions. This is very 'rough', but it should do the job. The payload is a
dictionary that stores named payload items for consumption by extensions.

You can use this to provide extensions that perform post-processing of result
data and storage of result data in the database/a file or wherever.


Former-commit-id: 15b82b0971600d8114e16ff91d56896139d4e11e
Former-commit-id: e032542bd592c29fcf5b835f0bddd3249501ae68 [formerly a66f67edc6d37a72c755be13ff52da887269ae4c] [formerly 21599b0d61138f9dc14973edbd763b888907f4c8 [formerly f0c1b77296ed0cc051aeb2e696ef07333cbf7cd0]]
Former-commit-id: 5143c79868bc04fbeabf0e3b37f8129cf2b3265b [formerly 59078c70da1e519a9ce97870ccb0d487e3c0081d]
Former-commit-id: 4415ff5c4c375096cacd31da1a9f6c3a47cb93c9