extensions: add payload support
This adds support for custom payload to communicate run results between
extensions. This is very 'rough', but it should do the job. The payload is a
dictionary that stores named payload items for consumption by extensions.

You can use this to provide extensions that perform post-processing of result
data and storage of result data in the database/a file or wherever.