project: replace prepare/download/build with a single compile method
This method can and will be auto-generated when a project uses
with_git or with_wget decorators.

The reduced interface mirrors the two modes of operation in benchbuild:

Compilation-Only experiments or
Runtime experiments.

The two corresponding methods a project has to support are called:
  compile(..)
  and run_tests(..)

For an easy transition to the new API you should just copy/paste the
code in the following order:
 compile(..):
	prepare()
     download()
     build()

This should produce the same result as before.


Former-commit-id: 3f1f59fbb4b86aceebe050510fbadcfb8b7d93fa
Former-commit-id: e0265f4281323759cdf6215b8d7abc29b0efc8e2 [formerly c2357706958a82345eec5beeb0a0ef01514e6a74] [formerly b6cf2efc857796a6ac26b7532ef33d4778d89b66 [formerly aaf0f3d0be0907bd042347405195559800cc50b0]]
Former-commit-id: 3b123baafb19fae766bb2d55482a3fe793e50e43 [formerly cf9cbcf9391b6839a6ef1f8bced48bad51fdfc87]
Former-commit-id: 599be0a4ce711edd6d2d2d1300e3b5b906cd9398