project: replace prepare/download/build with a single compile method
This method can and will be auto-generated when a project uses
with_git or with_wget decorators.

The reduced interface mirrors the two modes of operation in benchbuild:

Compilation-Only experiments or
Runtime experiments.

The two corresponding methods a project has to support are called:
  compile(..)
  and run_tests(..)

For an easy transition to the new API you should just copy/paste the
code in the following order:
 compile(..):
	prepare()
     download()
     build()

This should produce the same result as before.