settings: generalize 'env' section
This allows a user to customize the complete environment of benchbuild commands.
The 'env' section in the configuration is now capable of any environment
variables, instead of PATH and LD_LIBRARY_PATH only.

So far, benchbuild will still only take special care of PATH and
LD_LIBRARY_PATH, but this will change to a general solution in a future commit.


Former-commit-id: 553e491147970e8a4c19f5985e0b84c59c6fc44d
Former-commit-id: beee1d5fc43f6729ea21fd9753f25099b43dee14 [formerly d5e36785a939ac5229fc4615d10b021155b6023c] [formerly d233164818deef7aa76f47bc679af6a62c554860 [formerly 3c4ef459d845c393b87f3eeb875dde22d86e7be4]]
Former-commit-id: 16f9b2e19737c6dc4f77420bcecc531d4373d9db [formerly 65de5966544bf12a2915fe2dd35e8777d1e81a8f]
Former-commit-id: 6bc10edeeb67f2c52aa0feae005d305d571cb703