settings: generalize 'env' section
This allows a user to customize the complete environment of benchbuild commands.
The 'env' section in the configuration is now capable of any environment
variables, instead of PATH and LD_LIBRARY_PATH only.

So far, benchbuild will still only take special care of PATH and
LD_LIBRARY_PATH, but this will change to a general solution in a future commit.