gentoo: rework gentoo handling inside benchbuild
This is a major rebuild of all gentoo based projects inside benchbuild.
Before, we would treat gentoo projects like ordinary projects with the twist of
an uchroot container in the mix.
This forced project authors to deeply think about the folder structure and the
configuration of uchroot. Furthermore, it complicated handling of compiler and
binary substitution, warranting special implementations for both in the presence
of uchroot containers.

With this rework, we change that. Projects can declare a container they want to
use (e.g., the Gentoo container). From there, benchbuild will detect the need
for a container and continue the experiment inside a new benchbuild instance
running inside the container.
The configuration, mounts and prefixes will be redirected into the container as
well.

This greatly simplifies project handling for developers. You can now write
containerized projects in the same way as normal projects.

For further assistance you can use the container tool to inspect an unpacked
container.

A future commit will add detailed documentation on all the nitty gritty details
of container handling inside benchbuild.


Former-commit-id: 4767f9bf8ddf7ccbe3cb37f67962bb83079ac3c6
Former-commit-id: 2657a3723f0612f77cb482e0c088e28ed7fff93b [formerly 8774d62f8888d9bcae99484218635ece4d84d82d] [formerly 55fdd6418960e49ca05481a6b5468abba3b92b16 [formerly 2e1b725f2e0bb108b687d19b6315724c531e3f27]]
Former-commit-id: a362b1f505f9f383a997654a276257c4f622eab8 [formerly 4f929e9f858df94764648e85a6068e28ec418ddb]
Former-commit-id: b77a456510560df1b4093e211a01104023251325