gentoo: rework gentoo handling inside benchbuild
This is a major rebuild of all gentoo based projects inside benchbuild.
Before, we would treat gentoo projects like ordinary projects with the twist of
an uchroot container in the mix.
This forced project authors to deeply think about the folder structure and the
configuration of uchroot. Furthermore, it complicated handling of compiler and
binary substitution, warranting special implementations for both in the presence
of uchroot containers.

With this rework, we change that. Projects can declare a container they want to
use (e.g., the Gentoo container). From there, benchbuild will detect the need
for a container and continue the experiment inside a new benchbuild instance
running inside the container.
The configuration, mounts and prefixes will be redirected into the container as
well.

This greatly simplifies project handling for developers. You can now write
containerized projects in the same way as normal projects.

For further assistance you can use the container tool to inspect an unpacked
container.

A future commit will add detailed documentation on all the nitty gritty details
of container handling inside benchbuild.