settings: make value() a property and add conversion functions.
This converts the generic accessor 'value' of a configuration object from a
function to a property.

In addition, we add conversion functions for the common types bool and int.
Furthermore, appending to list values is now supported by configuration objects.

CFG['list_option'] += a
Now appends a to the list value, if there is one.