settings: make value() a property and add conversion functions.
This converts the generic accessor 'value' of a configuration object from a
function to a property.

In addition, we add conversion functions for the common types bool and int.
Furthermore, appending to list values is now supported by configuration objects.

CFG['list_option'] += a
Now appends a to the list value, if there is one.


Former-commit-id: 6c69bfddd4b581ca05b1e008234c31c95c714f80
Former-commit-id: d12f70e003c2b528d14f9eb8f9429f3905ccda21 [formerly 4e52f7b40fbe5ca1326ea6cdf3d7650c4eb0ddf4] [formerly 4097c28f55ee643163052e0b00e6a9c98800f468 [formerly 0da59916b72ef5aa12ef35c368e40bd07c2db6f0]]
Former-commit-id: d4f40159d9bc5c13d12eeb504c60ac4d30a05b1d [formerly 224056fb4eb305b488bd5aa7ea79ebbff0862bda]
Former-commit-id: 6574eecaf47e4535a03ded87f87b5066fe926bfe