portage_gen: use container version as base version for generated gentoo projects
This speeds up version generation for gentoo projects, if they are generated via
PortageFactory.

It doesn't really make sense to download a complete container, unpack it and
parse the emerge pretend output just to get the version of a gentoo package.
In a future series of patches, we might think about an easier way of handling
containers (docker maybe?).