download: replace source_required with clone_needed
This avoids hashing a complete git working directory to generate a hash.
This will break with every feature that works with the git repository in
any way and resort to redownloading everything.

clone_needed: this new check verifies that the repository that is
required matches with the origin remote url of an existing .git
directory. This results in the same semantics as before, without a
dependency on the directory contents.

This should create a more fluid experience with git worktree and avoid
redownloads that lead to data races on SLURM.

However, we still need to take care of a critical section in the
action sequence:
  git worktree prune
  git worktree add

This needs to be atomic accross multiple processes.