refactor: replace utils.run.run [breaks: <=4.0.1]
Benchbuild so far recognizes two different experimentation modes: compilation
and runtime.
This limitation is quite arbitrary and with this commit we begin to lift this
limitation and provide the user more control over the experimentation "phases".

As a first step it is necessary to unify the API used for both phases.
Historically, we provided a 'runner' argument that redirected all program output
to benchbuild's stdout.
The compilation stage did not have this and simply imported the
benchbuild.utils.run::run function and used it directly (this was the same
function that was handed in by benchbuild ;-)).

With future commits in mind, we replace benchbuild.utils.run::run with a new
function wrapper that takes a command and returns a function that takes the
command arguments and desired return code (as before).
This removes a great deal of clutter from experiments that make heavy use of the
former 'runner' argument.

However, this requires a mechanical change to existing code. I decided to use a
hard break, because the process is straight-forward to implement. Future changes
that depend on it might(will) break in subtle ways, if we support the new and
the old way.

Changes necessary:

1. Check for the correct import, e.g.:
	 from benchbuild.utils.run import run

-> from bencubuild.utls import run

2. Change signature of <project>::run_tests.
   def run_tests(self, runner):

-> def run_tests(self):

3. Change usages of runner inside <project>::run_tests.
   runner(cmd["a","b"], retcode=1)

-> cmd_ = run.watch(cmd)
   cmd_("a", "b", retcode=1)

4. Change usages of the benchbuild.utils.run.run function
   run(cmd["a","b"], retcode=1)

-> cmd_ = run.watch(cmd)
   cmd_("a", "b", retcode=1)

These changes are fairly mechanical. However, we cannot easily determine the
correct locations to provide an automatic fix.