containers: add cli subcommand
this builds the initial container image per project version.
We still need to consider the image-id or some kind of versioning scheme,
if we want to achieve reproducible builds.

One way to achieve this would be:
  One image per project/experiment version (!). With this we could version
  the image as well.

However, for now we just want to capture the build environment that is necessary
to experiment on the project.

Squash:
 - containers: cleaner implementation of declarative container support